+++ Assets/Gemserk.UPMGitPusher/Editor/Preferences.cs
@@ -7,6 +7,7 @@ namespace Gemserk.UPMGitPusher.Editor
         public const string PreferenceKeyDryRun = "UPMGitPusher.DryRun";
         public const string PreferenceKeyAutoCommit = "UPMGitPusher.AutoCommit";
         public const string PreferenceKeyCommitMessage = "UPMGitPusher.CommitMessage";
+        public const string PreferenceKeyAutoSearchPackages = "UPMGitPusher.AutoSearchPackages";
         
         // TODO: allow configure remote (default is origin)
         // TODO: optional configure path to package.json or reference to text asset itself
@@ -14,8 +15,10 @@ namespace Gemserk.UPMGitPusher.Editor
         public static bool dryRun => EditorPrefs.GetBool(PreferenceKeyDryRun, false);
         public static bool automaticCommit => EditorPrefs.GetBool(PreferenceKeyAutoCommit, true);
 
-        public static string commitmMessage => EditorPrefs.GetString(PreferenceKeyCommitMessage, 
+        public static string commitMessage => EditorPrefs.GetString(PreferenceKeyCommitMessage, 
             "Updated version from {PREVIOUS_VERSION} to {NEW_VERSION}");
+        
+        public static bool AutoSearchPackages => EditorPrefs.GetBool(PreferenceKeyAutoSearchPackages, false);
 
     }
 }