+++ Assets/Gemserk.UPMGitPusher/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Changed to search for a list of packages to export.
 - Export only selected package.json if file selected in editor.
+- Toggleable automatic package search (disabled by default)
 
 # 0.0.18